+++ spotify-service/src/main/java/com/pinapp/spotifyservice/controller/ArtistController.java
@@ -1,8 +1,8 @@
 package com.pinapp.spotifyservice.controller;
 
 import com.pinapp.spotifyservice.controller.request.ArtistRequest;
-import com.pinapp.spotifyservice.domain.Artist;
-import com.pinapp.spotifyservice.service.Implementations.ArtistService;
+import com.pinapp.spotifyservice.domain.model.Artist;
+import com.pinapp.spotifyservice.service.Implementation.ArtistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
