+++ spotify-service/src/main/java/com/pinapp/spotifyservice/controller/SpotifyController.java
@@ -1,7 +1,7 @@
 package com.pinapp.spotifyservice.controller;
 
-import com.pinapp.spotifyservice.domain.Track;
-import com.pinapp.spotifyservice.service.Implementations.SpotifyService;
+import com.pinapp.spotifyservice.domain.model.Track;
+import com.pinapp.spotifyservice.service.Implementation.SpotifyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
