+++ spotify-service/src/main/java/com/pinapp/spotifyservice/domain/model/Album.java
@@ -1,4 +1,4 @@
-package com.pinapp.spotifyservice.domain;
+package com.pinapp.spotifyservice.domain.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,12 +6,11 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Artist {
+@Builder
+public class Album {
+  private Long idAlbum;
   private Long idArtist;
   private String name;
-  private String genre;
-  private String image;
 }
