+++ spotify-service/src/main/java/com/pinapp/spotifyservice/service/IArtistService.java
@@ -1,7 +1,7 @@
 package com.pinapp.spotifyservice.service;
 
 import com.pinapp.spotifyservice.controller.request.ArtistRequest;
-import com.pinapp.spotifyservice.domain.Artist;
+import com.pinapp.spotifyservice.domain.model.Artist;
 
 import java.util.List;
