+++ spotify-service/src/main/java/com/pinapp/spotifyservice/service/ISpotifyService.java
@@ -1,6 +1,6 @@
 package com.pinapp.spotifyservice.service;
 
-import com.pinapp.spotifyservice.domain.Track;
+import com.pinapp.spotifyservice.domain.model.Track;
 
 public interface ISpotifyService {
   Track playTrack(Long id);
