+++ spotify-service/src/main/java/com/pinapp/spotifyservice/service/Implementation/TrackService.java
@@ -1,8 +1,8 @@
-package com.pinapp.spotifyservice.service.Implementations;
+package com.pinapp.spotifyservice.service.Implementation;
 
 import com.pinapp.spotifyservice.controller.request.TrackRequest;
-import com.pinapp.spotifyservice.domain.Track;
-import com.pinapp.spotifyservice.domain.mappers.TrackMapper;
+import com.pinapp.spotifyservice.domain.model.Track;
+import com.pinapp.spotifyservice.domain.mapper.TrackMapper;
 import com.pinapp.spotifyservice.service.ITrackService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
